refactor: move FMEA view methods to metaclass
